backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
pypdf2==3.0.1
python-docx==0.8.11
python-multipart==0.0.5
requests==2.31.0
python-dotenv==0.21.1
urllib3==1.26.18
anthropic>=0.18.0
sqlalchemy>=2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.0
google-auth-oauthlib==1.1.0